# contracts/work-platform-bff-openapi.yaml
openapi: 3.0.3
info:
  title: Work Platform BFF API Contract (Agent-Lifeform)
  version: 1.0.0
  description: >
    工作平台（统一工作台）BFF 接口契约（W2 扩展）。
    覆盖 14 模块：总览/生命体征/决策沙盘/五感/进化/协作总线/任务/对话/专家/技能/连接器/自动化/灵感/审批。
    Mock 数据必须与该契约结构完全一致 → 切换零 UI 改动（R-C09 延续）。
    设计依据：《工作平台功能设计文档》v1.0 §6。

    **统一错误信封（2026-09-18 起全平台一致）**：所有错误响应均为
    `{code, message, details}`，与 gateway-service、Java 三服务、nlp-service 完全同构；
    BFF 路由层分流：路径不存在 → 404 `AGENT_NOT_FOUND`，路径存在但方法不支持 →
    405 `AGENT_METHOD_NOT_ALLOWED`（附 `Allow` 头），请求体非法 → 400 `AGENT_BAD_REQUEST`，
    鉴权失败 → 401 `AGENT_UNAUTHORIZED` / 403 `AGENT_FORBIDDEN`。
    区分 404 与 5xx 是前端降级判定的前提（404 = 端点未实现 = planned）。
    完整异常传播链见 `docs/异常流程归纳.md` §2。
servers:
  - url: /api/wp
    description: Work Platform BFF 聚合层
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ErrorEnvelope:
      description: 统一错误响应（`{code, message, details}`）—— 见 info.description
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    ApiResponse:
      type: object
      required: [code, data, msg]
      properties:
        code: { type: integer, example: 0 }
        data: { type: object }
        msg: { type: string, example: ok }
    ErrorEnvelope:
      type: object
      description: |
        全平台统一错误信封。2026-09-18 起 BFF（原 `{error}`）与 nlp-service
        （原 FastAPI 默认 `{detail}`）已对齐到本结构，客户端只需一套解析逻辑。
      required: [code, message, details]
      properties:
        code:
          type: string
          description: 统一错误码（R1-08）
          enum:
            - AGENT_BAD_REQUEST
            - AGENT_UNAUTHORIZED
            - AGENT_FORBIDDEN
            - AGENT_NOT_FOUND
            - AGENT_METHOD_NOT_ALLOWED
            - AGENT_CONFLICT
            - AGENT_DUPLICATE
            - AGENT_TIMEOUT
            - AGENT_BUS_UNAVAILABLE
            - AGENT_UPSTREAM_UNAVAILABLE
            - AGENT_INTERNAL_ERROR
            # Phase 5 四肢期：工具域专属码（tool-executor `ErrorCode` 同源副本）
            # 「被拒 ≠ 服务不可用」：前四个码都是**已落审计的真实调用结果**，
            # 出现在 200 + success:false 的执行响应里，而不是 5xx。
            - AGENT_TOOL_NOT_ALLOWED
            - AGENT_TOOL_ARGS_INVALID
            - AGENT_TOOL_ARGS_BLOCKED
            - AGENT_TOOL_EXEC_FAILED
            - AGENT_TOOL_CIRCUIT_OPEN
            - AGENT_SANDBOX_REJECTED
          example: AGENT_NOT_FOUND
        message: { type: string, example: 接口不存在：/api/wp/not-exist }
        details:
          type: object
          description: 上下文明细（字段级原因 / method+supported / path 等）
          additionalProperties: true
          example: { path: /api/wp/not-exist }
...
paths:
...
  /tools:
    x-wp-status: implemented
    get:
      summary: 'R-C05(预) 执行视图聚合：工具清单 / 注册表 / 指标 / 沙箱 / 审计（代理 tool-executor）'
      description: >-
        tool-executor 不可用时回 200 + available=false + reason + reason_code（降级可见，非 5xx），
        分片缺失时回 partial=true + partial_reasons + partial_note（不把「取不到」写成 0）。
        **降级码三分**：`*_unavailable`（对端未启动/连接被拒）、`*_timeout`（对端未在预算内返回，
        是「慢」不是「没有」）、`*_missing`（端点 404，对端版本旧于本 BFF）。
        各类预算见响应字段 `probe_budget_ms`（沙箱端点需探测 Docker，预算单独放宽）。
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '405':
          $ref: '#/components/responses/ErrorEnvelope'
  /tools/registry:
    x-wp-status: implemented
    get:
      summary: 'IN-06 工具注册表摘要 + 变更历史（semver 与废弃期策略）'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '405':
          $ref: '#/components/responses/ErrorEnvelope'
  /tools/audit:
    x-wp-status: implemented
    get:
      summary: 'R5-08 工具调用审计明细（租户隔离，最近 N 条）'
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, default: 50 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '405':
          $ref: '#/components/responses/ErrorEnvelope'
  /tools/audit/stats:
    x-wp-status: implemented
    get:
      summary: 'R5-08 工具调用审计汇总（总数 / 成功 / 拦截 / 沙箱化）'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '405':
          $ref: '#/components/responses/ErrorEnvelope'
...
  /tools/sandbox:
    x-wp-status: implemented
    get:
      summary: 'R-C05(预) 沙箱状态：生效后端 / 是否真隔离 / 降级原因'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '405':
          $ref: '#/components/responses/ErrorEnvelope'
  /tools/execute:
    x-wp-status: implemented
    post:
      summary: 'R5-02 执行工具（控制端点：来源白名单 + X-WP-Control-Token）'
      description: >-
        工具被守卫/参数校验/沙箱拒绝时回 200 + available=true + success=false + error_code
        （已落审计的真实调用结果）；仅当四肢层不可达时回 available=false + reason。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tool_name]
              properties:
                tool_name: { type: string }
                arguments: { type: object }
                call_id: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/ErrorEnvelope'
        '403':
          $ref: '#/components/responses/ErrorEnvelope'
        '405':
          $ref: '#/components/responses/ErrorEnvelope'
  /tools/{name}:
    x-wp-status: implemented
    get:
      summary: '单工具详情（含 JSON Schema 指纹、超时、沙箱要求与熔断状态）'
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '405':
          $ref: '#/components/responses/ErrorEnvelope'
  /tools/{name}/impact:
    x-wp-status: implemented
    get:
      summary: 'IN-06 工具变更影响分析（受影响 Agent / 流程 / 是否需契约测试）'
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '405':
          $ref: '#/components/responses/ErrorEnvelope'
...